Move 'DynamicFieldsModelSerializer' class from 'loans.serializers' to 'codenation.utils.serializers'
